requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
attrs==25.4.0
babel==2.17.0
bidict==0.23.1
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
courlan==1.3.2
dateparser==1.2.2
distro==1.9.0
docutils==0.22.3
exceptiongroup==1.3.1
fastapi==0.124.4
feedparser==6.0.12
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
htmldate==1.9.4
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
idna==3.11
ifaddr==0.2.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
jusText==3.0.2
lxml==6.0.2
lxml_html_clean==0.4.3
macholib==1.16.4
markdown2==2.5.4
MarkupSafe==3.0.3
multidict==6.7.0
nicegui==3.4.1
openai==2.13.0
orjson==3.11.5
packaging==25.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.12.3
python-multipart==0.0.20
python-socketio==5.15.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
sgmllib3k==1.0.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
tld==0.13.1
tqdm==4.67.1
trafilatura==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.2
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
wsproto==1.2.0
yarl==1.22.0
zipp==3.23.0
